Initial version of LoadingController
Changes:
- Loading logic moved from LoadVehicleGarage to new LoadingController
- Calling classes updated
- LoadVehicleGarage constructors are consolidated to use the same logic.
#439
