+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/person/ai/mission/RoverMission.java
@@ -365,9 +365,7 @@ public abstract class RoverMission extends VehicleMission {
 								if (!hasAnotherMission && isRoverInAGarage) {
 
 									assignTask(person,
-												new LoadVehicleGarage(person, v,
-														getRequiredResourcesToLoad(), getOptionalResourcesToLoad(),
-														getRequiredEquipmentToLoad(), getOptionalEquipmentToLoad()));
+												new LoadVehicleGarage(person, this));
 								} else {
 									// Check if it is day time.
 									assignTask(person, new LoadVehicleEVA(person, v,
