+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/person/ai/mission/Trade.java
@@ -534,9 +534,7 @@ public class Trade extends RoverMission implements Serializable {
 							if (member instanceof Person) {
 								Person person = (Person) member;
 								assignTask(person,
-										new LoadVehicleGarage(person, getVehicle(), getRequiredResourcesToLoad(),
-												getOptionalResourcesToLoad(), getRequiredEquipmentToLoad(),
-												getOptionalEquipmentToLoad()));
+										new LoadVehicleGarage(person, this));
 							}
 						} else {
 							if (member instanceof Person) {
